refactor: reorganize violin guard core, unify locking and json storage, optimize facades
